feat: move all pages to app folder, add splashscreen
