+++ apps/desktop/src-tauri/capabilities/default.json
@@ -2,7 +2,7 @@
 	"$schema": "../gen/schemas/desktop-schema.json",
 	"identifier": "default",
 	"description": "Capability for the main window",
-	"windows": ["main*"],
+	"windows": ["main*", "splashscreen"],
 	"permissions": [
 		{
 			"identifier": "http:default",
