+++ apps/desktop/src-tauri/tauri.macos.conf.json
@@ -4,9 +4,16 @@
 		"windows": [
 			{
 				"hiddenTitle": true,
+				"url": "/app",
 				"title": "Kunkun",
 				"width": 800,
+				"visible": false,
 				"height": 600
+			},
+			{
+				"url": "/splashscreen",
+				"visible": false,
+				"label": "splashscreen"
 			}
 		]
 	}
