+++ apps/desktop/src/lib/components/standalone/settings/AddDevExtForm.svelte
@@ -87,7 +87,7 @@
 	async function pickExtFiles() {
 		if (!$appConfig.devExtensionPath) {
 			toast.warning("Please set the dev extension path in the settings")
-			return goto("/settings/set-dev-ext-path")
+			return goto("/appsettings/set-dev-ext-path")
 		}
 		const selected = await openFileSelector({
 			directory: false,
