+++ apps/desktop/src/lib/utils/key.ts
@@ -76,7 +76,7 @@ export async function globalKeyDownHandler(e: KeyboardEvent) {
 	if ((_platform === "macos" && e.metaKey) || (_platform === "windows" && e.ctrlKey)) {
 		if (e.key === ",") {
 			e.preventDefault()
-			goto("/settings")
+			goto("/appsettings")
 		}
 	}
 	// Toggle Devtools with control + shift + I
